Move MainActivity in main package
